TailwindCSS and Svelte Stellar (#1666)
* Removes the bootstrap utility css and swaps a few of those classes over to TailwindCSS classes.
* Updates typescript to 5.7.3
* This PR uses the new Stellar Svelte components and TailwindCSS classes in a variety of pages and components. The following is a list of changed pages and components:
* Plans page reworked with new components and classes
* ContextMenu -> Stellar ContextMenu and/or DropdownMenu (affected a variety of pages and components)
* Various icon changes to use Lucide Icons, most notably in the app menu
* App navbar reworked with new components and classes. Additionally changed relevant menu items to be links for better UX.
* Field and FieldError lightly reworked to use new classes
* Login page lightly reworked to use new components and classes
* About modal lightly reworked to use new classes
* ModelStatusRollup reworked to use new components and classes
* ParameterInfo reworked to use new components and classes
* DatePicker lightly reworked to use new components and classes. Improved accessibility of the component but still work to go, could eventually try using the Stellar calendar component to start with which would bring better keyboard nav.

+++ src/components/ui/DataGrid/SingleActionDataGrid.svelte
@@ -8,16 +8,15 @@
     editItem: CustomEvent<RowId[]>;
   }
   import { browser } from '$app/environment';
+  import { ContextMenu } from '@nasa-jpl/stellar-svelte';
   import type { ColDef, ColumnState, IRowNode, RedrawRowsParams } from 'ag-grid-community';
   import { createEventDispatcher, onDestroy, type ComponentEvents } from 'svelte';
   import type { User } from '../../../types/app';
   import type { Dispatcher } from '../../../types/component';
-  import type { RowId, TRowData } from '../../../types/data-grid';
+  import type { RowId } from '../../../types/data-grid';
   import type { PermissionCheck } from '../../../types/permissions';
   import { isDeleteEvent } from '../../../utilities/keyboardEvents';
   import { permissionHandler } from '../../../utilities/permissionHandler';
-  import ContextMenuHeader from '../../context-menu/ContextMenuHeader.svelte';
-  import ContextMenuItem from '../../context-menu/ContextMenuItem.svelte';
   import DataGrid from '../../ui/DataGrid/DataGrid.svelte';
 
   const defaultDeletePermissionError: string = 'You do not have permission to delete.';
@@ -161,45 +160,30 @@
   on:selectionChanged
 >
   <svelte:fragment slot="context-menu">
-    <ContextMenuHeader>Actions</ContextMenuHeader>
     {#if hasEdit}
-      <ContextMenuItem
-        use={[
-          [
-            permissionHandler,
-            {
-              hasPermission: editPermission,
-              permissionError: editPermissionError,
-            },
-          ],
-        ]}
-        on:click={editItem}
+      <div
+        use:permissionHandler={{
+          hasPermission: editPermission,
+          permissionError: editPermissionError,
+        }}
       >
-        Edit {itemDisplayText}
-      </ContextMenuItem>
+        <ContextMenu.Item size="sm" disabled={!editPermission} on:click={editItem}>
+          Edit {itemDisplayText}
+        </ContextMenu.Item>
+      </div>
     {/if}
     {#if selectedItemId !== null}
-      <ContextMenuItem
-        use={[
-          [
-            permissionHandler,
-            {
-              hasPermission: deletePermission,
-              permissionError: deletePermissionError,
-            },
-          ],
-        ]}
-        on:click={deleteItem}
+      <div
+        use:permissionHandler={{
+          hasPermission: deletePermission,
+          permissionError: deletePermissionError,
+        }}
       >
-        Delete {itemDisplayText}
-      </ContextMenuItem>
+        <ContextMenu.Item size="sm" disabled={!deletePermission} on:click={deleteItem}>
+          Delete {itemDisplayText}
+        </ContextMenu.Item>
+      </div>
     {/if}
+    <ContextMenu.Separator />
   </svelte:fragment>
 </DataGrid>
-
-<style>
-  :global(.context-menu-item.disabled) {
-    cursor: not-allowed;
-    opacity: 0.5;
-  }
-</style>
